[clDNN][Build] Quick fix for warning.

+++ tests/test_cases/softmax_gpu_test.cpp
@@ -68,7 +68,7 @@ public:
 
 TEST_F(softmax_gpu_xb_f32_test_fixture, input_same_values) {
 // in_buffer filled with same value == 1.0f
-    for(size_t i = 0; i < out_size; ++i) {
+    for(uint32_t i = 0; i < out_size; ++i) {
               in_buffer[i] = 1.0f;
         expected_buffer[i] = 0.1f;
     }
@@ -81,7 +81,7 @@ TEST_F(softmax_gpu_xb_f32_test_fixture, input_same_values) {
     execute({input, output, act}).wait();
 
     auto output_ptr = output.as<const memory&>().pointer<float>();
-    for (size_t i = 0; i < out_size; i++)
+    for (uint32_t i = 0; i < out_size; i++)
     {
         out_buffer[i] = get_value<float>(output_ptr, i);
     }
@@ -106,7 +106,7 @@ TEST_F(softmax_gpu_xb_f32_test_fixture, input_same_values_batch_wise) {
     execute({input, output, act}).wait();
 
     auto output_ptr = output.as<const memory&>().pointer<float>();
-    for (size_t i = 0; i < out_size; i++)
+    for (uint32_t i = 0; i < out_size; i++)
     {
         out_buffer[i] = get_value<float>(output_ptr, i);
     }
@@ -155,7 +155,7 @@ TEST_F(softmax_gpu_xb_f32_test_fixture, values_batch_wise) {
 
     execute({input, output, act}).wait();
     auto output_ptr = output.as<const memory&>().pointer<float>();
-    for (size_t i = 0; i < out_size; i++)
+    for (uint32_t i = 0; i < out_size; i++)
     {
         out_buffer[i] = get_value<float>(output_ptr, i);
     }
